Added custom treatment for exceptions

+++ event-ticket-api/src/main/java/com/example/booking/controller/request/ticket/EmmitTicketRequest.java
@@ -10,5 +10,5 @@ public record EmmitTicketRequest(
         UUID eventId,
 
         @NotNull(message = "Ticket category id is required!")
-        Integer ticketCategoryId
+        Long ticketCategoryId
 ) {}
