+++ event-ticket-api/src/main/java/com/example/booking/controller/response/ticket/CreateTicketResponse.java
@@ -2,5 +2,5 @@ package com.example.booking.controller.response.ticket;
 
 import java.util.UUID;
 
-public record CreateTicketResponse(UUID ticketId, UUID eventId, UUID userId, Integer ticketCategoryId) {
+public record CreateTicketResponse(UUID ticketId, UUID eventId, UUID userId, Long ticketCategoryId) {
 }
